+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * **(breaking-change)** Rename plugin to RedmineRecurringTasks
 * **(breaking-change)** Transfer plugin to southbridge (http://github.com/centosadmin)
+* Make preserve author of issue on copy
 
 # 0.0.1
 * Add ability to make issues copies